Upgraded to spring-security-oauth2 2.0.10.release and refactor oauth security to separate java class
